[gce-saltcloud] Moving write-to-log to utils.
